requirements.txt
appium-python-client==2.6.1
selenium==4.9.0
urllib3<2
pytest
allure-pytest
python-dotenv
